Fix custom avatar ground offset
This commit addresses two original pieces of code;
1. Instead of using the default ("og") player model's hip vertical (y) offset for ground calculations it will use the loaded avatar for calculating the offset.
2. Referring to the toe bones of the avatar instead of the feet bones. The reason for this is because the feet "bones" positions refer to the "head"s of that bone instead of the "tail"s, resulting in a smaller offset than intended. Referring instead to the toe bones will refer to something closer to what was intended. Still not a -perfect- solution, but it'll work in more cases than what existed prior.

One problem that has not been addressed by this patch is that if there are no existing toe bones on an avatar, that there will be a null reference which might (and probably will) break some avatars, these situations can be possible since toe bones are not a requirement for humanoid rig definitions.- There needs to be a fallback case in those fringe situations.

Besides that, currently we're referring to the toe bone by taking the first child of the foot bones.- I'm sure there's a better way to address this too but it's not as much of an issue in comparison to the previous addressed point.